restored summary.js for testing purposes, commented out summary.ts
